+++ xinu/freemem.c
@@ -6,26 +6,26 @@
 //  freemem  --  free a memory block, returning it to memlist
 //------------------------------------------------------------------------
 SYSCALL
-freemem(void *b, unsigned size)
+freemem(void *b, size_t size)
 {
 	int ps;
 	struct mblock *p, *q, *block = (struct mblock *)b;
 	unsigned top;
 
-	if (size == 0 || (unsigned) block > (unsigned) maxaddr
-	    || ((unsigned) block) < ((unsigned) &end))
+	if (size == 0 || (u32)block > (u32)maxaddr || ((u32)block) < ((u32)&end))
 		return SYSERR;
-	size = (unsigned) roundew(size);
+	size = (u32)roundew(size);
 	ps = disable();
 	for (p = memlist.mnext, q = &memlist;
-	     (char *)p != NULL && p < block; q = p, p = p->mnext);
-	if ((top = q->mlen + (unsigned) q) > (unsigned) block
-	    && (q != &memlist || (char *)p != NULL)
-	    && (size + (unsigned) block) > (unsigned) p) {
+	     (char *)p != NULL && p < block;
+	     q = p, p = p->mnext);
+	if ((top = q->mlen + (u32)q) > (u32)block &&
+	    (q != &memlist || (char *)p != NULL) &&
+	    (size + (u32)block) > (u32)p) {
 		restore(ps);
 		return SYSERR;
 	}
-	if (q != &memlist && top == (unsigned) block)
+	if (q != &memlist && top == (u32)block)
 		q->mlen += size;
 	else {
 		block->mlen = size;
@@ -33,10 +33,11 @@ freemem(void *b, unsigned size)
 		q->mnext = block;
 		q = block;
 	}
-	if ((unsigned) (q->mlen + (unsigned) q) == (unsigned) p) {
+	if ((u32)(q->mlen + (u32)q) == (u32)p) {
 		q->mlen += p->mlen;
 		q->mnext = p->mnext;
 	}
 	restore(ps);
+
 	return OK;
 }
